BUGFIX: Prevent not fully translated content from getting lost in move operations
This resolves the issue that moving a collection with subnodes in dimension value A doesn’t update
the parent path of nodes in dimension value B if the collection itself does not exist in language B. The result would be that the content in B is now disconnected from the tree and not reachable anymore.

This slightly changes the behavior in existing projects, but the change keeps content from getting lost and is therefore essential for every Neos project with dimensions and fallbacks.

Resolves: #3340